Feature/radius slider filter

+++ app/src/main/java/com/github/se/icebreakrr/model/profile/ProfilesRepository.kt
@@ -20,7 +20,7 @@ interface ProfilesRepository {
 
   fun getProfilesInRadius(
       center: GeoPoint,
-      radiusInMeters: Double,
+      radiusInMeters: Int,
       onSuccess: (List<Profile>) -> Unit,
       onFailure: (Exception) -> Unit
   )
